Retrieve Nitrogen deps after generate
This is instead of predownloading and compiling them.  This is to assist
the sync application, which uses the path names from
Module:module_info(compile) to determine the source file. In a forked
version of sync specifically for Nitrogen, this will then allow sync to
properly infer the *new* location of files after a compiled release has
been moved to a new directory or a new server altogether.