Generate rebar.config for each platform
This is instead of having a *slightly* different copy of rebar.config in
each overlay for each platform. This will vastly simplify the release
process, requiring only a few changes rather than having to update 5
different rebar.config files for each new official release.

Also cleaned up the Makefile to limit the copy-and-paste for each
platform.