increase default max POST chunk size for Yaws
The default max chunk size for POST data for Yaws before it breaks the data
into chunks is 10240 bytes. Increase this to 2048000 for now because
currently simple_bridge doesn't handle the chunking Yaws does for POST data
that exceeds the default size. Note that this doesn't fix the problem, but
is just an workaround that makes it less likely users will hit it.

Unfortunately I don't have an automated test for this, but I verified it
manually. I'd rather put testing effort into a proper fix in simple_bridge.